requirements.txt
ultralytics~=8.3.13
pillow~=10.4.0
torch~=2.4.1+cu118